Refactoring (extracting common functions)
