+++ countryParser.py
@@ -1,10 +1,12 @@
+from common import *
+
 def main(fileName):
     global outputDict
-    inputFile = structureFile("history/countries/"+fileName) #Transcribes game file to more parseable format
+    inputFile = structureFile(path, "history/countries/", fileName) #Transcribes game file to more parseable format
     outputDict = {}
     for line in inputFile:
         #Determines how deeply nested the current line is
-        nestingCheck(line)
+        nesting, nestingIncrement = nestingCheck(line, nesting, nestingIncrement)
         if nesting > 0:
             continue
         command, value = getValues(line)
@@ -29,86 +31,6 @@ def main(fileName):
     except UnboundLocalError:
         return ""
 
-#Reads in a statement file as a dictionary
-def readStatements(localisationName):
-    localisation = {}
-    with open('%s.txt' % localisationName) as effectsFile:
-        for line in effectsFile.readlines():
-            if not ":" in line: #Not a statement string
-                continue
-            StatementName, formatString = line.split(':', 1)
-            localisation[StatementName.strip()] = formatString.strip()
-
-    return localisation
-
-#Reads in a definition file as a dictionary
-def readDefinitions(name):
-    definitions = {}
-    with open(path+"/localisation/%s_l_english.yml" % name, encoding="utf-8") as definitionsFile:
-        lines = definitionsFile.readlines()
-        lineNumber = 1
-        for line in lines[1:]:
-            if not ":" in line:
-                continue
-            try:
-                identifier, value = line.split(':', 1)
-                definitions[identifier.strip()] = value.strip('" \n')
-            except ValueError:
-                print ("%d -> %s" % (lineNumber, line))
-            lineNumber += 1
-
-    return definitions
-
-linePattern = re.compile("(=[\s]*[\w\.]*) ([\w\.]*[\s]*=)")
-
-#Splits the file at every bracket to ensure proper parsing
-def structureFile(name):
-    functionOutput = []
-
-    with open('%s/%s' % (path, name), encoding="Windows-1252") as file:
-        for line in file:
-            if "#" in line:
-                line = line.split("#")[0]
-            if line == "":
-                continue
-            line = line.replace("{", "{\n").replace("}", "\n}").strip() #Splits line at brackets
-            if line == "":
-                continue
-            if "=" in line:
-                count = line.count("=")
-                if count > 1:
-                    for values in range(count):
-                        line = linePattern.sub("\g<1>\n\g<2>", line) #Splits lines with more than one statement in two
-            if "\n" in line:
-                parts = line.split("\n")
-                for p in parts:
-                    functionOutput.append(p)
-            else:
-                functionOutput.append(line)
-    return functionOutput
-
-#Determines the current level of nesting
-def nestingCheck(line):
-    global nesting
-    global nestingIncrement
-    nestingIncrement = 0
-    #Thanks to file restructuring, it is impossible for there to be multiple brackets on a line
-    if "{" in line:
-        nesting += 1
-        nestingIncrement = 1
-    elif "}" in line:
-        nesting -= 1
-        nestingIncrement = -1
-
-def getValues(line):
-    line = line.split("=")
-    line[0] = line[0].strip()
-    try: #Checks if the command has a value
-        line[1] = line[1].strip().strip('{}"')
-        return line[0], line[1]
-    except IndexError:
-        return line[0], ""
-
 def valueLookup(value):
     try:
         return provinces["PROV"+value]
@@ -128,78 +50,65 @@ def output(command, value): #Outputs line to a temp variable. Written to output 
     global outputDict
     outputDict[command] = value
 
-#Determines the current level of nesting
-def nestingCheck(line):
-    global nesting
-    global nestingIncrement
-    nestingIncrement = 0
-    #Thanks to file restructuring, it is impossible for there to be multiple brackets on a line
-    if "{" in line:
-        nesting += 1
-        nestingIncrement = 1
-    elif "}" in line:
-        nesting -= 1
-        nestingIncrement = -1
+if __name__ == "__main__":
+    import cProfile, pstats
+    pr = cProfile.Profile()
+    pr.enable()
 
-import cProfile, pstats
-pr = cProfile.Profile()
-pr.enable()
+    import time #Used for timing the parser
+    start = time.clock()
+    import os #Used to grab the list of files
+    settings = readStatements("settings")
+    path = settings["path"].replace("\\", "/")
 
-import time #Used for timing the parser
-start = time.clock()
-import re #Needed for various string handling
-import os #Used to grab the list of files
-settings = readStatements("settings")
-path = settings["path"].replace("\\", "/")
+    #Dictionaries of known statements
+    countryStatements = readStatements("statements/countryStatements")
+    countryCommands = countryStatements["commands"].split()
+    nesting, nestingIncrement = 0, 0
+    finalOutput = ""
+    #ideas = {}
+    #unformattedIdeas = structureFile(path, "common/ideas", "00_country_ideas.txt")
+    #triggerFound = False
+    #for line in unformattedIdeas:
+    #    nestingCheck(line)
+    #    if triggerFound:
+    #        if nesting == 1:
+    #            triggerFound = False
+    #        else:
+    #            ideas[token].append(line)
+    #    if nesting == 1 and nestingIncrement == 1:
+    #        token = getValues(line)[0]
+    #        ideas[token] = []
+    #    elif "trigger" in line:
+    #        triggerFound = True
 
-#Dictionaries of known statements
-countryStatements = readStatements("statements/countryStatements")
-countryCommands = countryStatements["commands"].split()
-nesting, nestingIncrement = 0, 0
-finalOutput = ""
-#ideas = {}
-#unformattedIdeas = structureFile("common/ideas/00_country_ideas.txt")
-#triggerFound = False
-#for line in unformattedIdeas:
-#    nestingCheck(line)
-#    if triggerFound:
-#        if nesting == 1:
-#            triggerFound = False
-#        else:
-#            ideas[token].append(line)
-#    if nesting == 1 and nestingIncrement == 1:
-#        token = getValues(line)[0]
-#        ideas[token] = []
-#    elif "trigger" in line:
-#        triggerFound = True
-
-try:
-    #Dictionaries of relevant values
-    provinces = readDefinitions("prov_names")
-    countries = readDefinitions("countries")
-    countries.update(readDefinitions("text"))
-    countries.update(readDefinitions("eu4"))
-    lookup = readDefinitions("eu4")
-    lookup.update(readDefinitions("text"))
-    lookup.update(readDefinitions("opinions"))
-    lookup.update(readDefinitions("powers_and_ideas"))
-    lookup.update(readDefinitions("decisions"))
-    lookup.update(readDefinitions("modifers"))
-    lookup.update(readDefinitions("muslim_dlc"))
-    lookup.update(readDefinitions("Purple_Phoenix"))
-    lookup.update(readDefinitions("core"))
-    lookup.update(readDefinitions("missions"))
-    lookup.update(readDefinitions("diplomacy"))
-    for fileName in os.listdir("%s/history/countries" % path):
-        print("Parsing file %s" % fileName)
-        finalOutput += main(fileName)
-    with open("output/countryOutput.txt", "w", encoding="utf-8") as outputFile:
-        outputFile.write(finalOutput)
-except FileNotFoundError:
-    print("File not found error: Make sure you've set the file path in settings.txt")
-elapsed = time.clock() - start
-print("Parsing the files took %.3f seconds" %elapsed)
-pr.disable()
-sortby = 'tottime'
-ps = pstats.Stats(pr).sort_stats(sortby)
-ps.print_stats()
+    try:
+        #Dictionaries of relevant values
+        provinces = readDefinitions(path, "prov_names")
+        countries = readDefinitions(path, "countries")
+        countries.update(readDefinitions(path, "text"))
+        countries.update(readDefinitions(path, "eu4"))
+        lookup = readDefinitions(path, "eu4")
+        lookup.update(readDefinitions(path, "text"))
+        lookup.update(readDefinitions(path, "opinions"))
+        lookup.update(readDefinitions(path, "powers_and_ideas"))
+        lookup.update(readDefinitions(path, "decisions"))
+        lookup.update(readDefinitions(path, "modifers"))
+        lookup.update(readDefinitions(path, "muslim_dlc"))
+        lookup.update(readDefinitions(path, "Purple_Phoenix"))
+        lookup.update(readDefinitions(path, "core"))
+        lookup.update(readDefinitions(path, "missions"))
+        lookup.update(readDefinitions(path, "diplomacy"))
+        for fileName in os.listdir("%s/history/countries" % path):
+            print("Parsing file %s" % fileName)
+            finalOutput += main(fileName)
+        with open("output/countryOutput.txt", "w", encoding="utf-8") as outputFile:
+            outputFile.write(finalOutput)
+    except FileNotFoundError:
+        print("File not found error: Make sure you've set the file path in settings.txt")
+    elapsed = time.clock() - start
+    print("Parsing the files took %.3f seconds" %elapsed)
+    pr.disable()
+    sortby = 'tottime'
+    ps = pstats.Stats(pr).sort_stats(sortby)
+    ps.print_stats()
