+++ parser.py
@@ -1,11 +1,15 @@
+from common import *
+
 def main(fileName):
     global outputText
-    inputFile = structureFile(fileName) #Transcribes game file to more parseable format
+    inputFile = structureFile(path, folder, fileName) #Transcribes game file to more parseable format
     specialSection, negative, negativeNesting, printSection, base_chance, option, random_list, randomNesting = False, False, False, False, False, False, False, False
     value1, value2, modifier, command, value = "", "", "", "", ""
     outputText = ""
     for line in inputFile:
-        nestingCheck(line) #Determines how deeply nested the current line is
+        global nesting
+        global nestingIncrement
+        nesting, nestingIncrement = nestingCheck(line, nesting, nestingIncrement) #Determines how deeply nested the current line is
         if nesting <= 1:
             continue #Nothing relevant is nested this low
         if nesting == 2:
@@ -147,94 +151,7 @@ def main(fileName):
             value2 = ""
     with open("output/%s" % fileName, "w", encoding="utf-8") as outputFile:
         outputFile.write(outputText)
- 
-#Reads in a statement file as a dictionary
-def readStatements(localisationName):
-    localisation = {}
-    with open('%s.txt' % localisationName) as effectsFile:
-        for line in effectsFile.readlines():
-            if not ":" in line: #Not a statement string
-                continue
-            StatementName, formatString = line.split(':', 1)
-            localisation[StatementName.strip()] = formatString.strip()
-         
-    return localisation
- 
-#Reads in a definition file as a dictionary
-def readDefinitions(name):
-    definitions = {}
-    with open(path+"/localisation/%s_l_english.yml" % name, encoding="utf-8") as definitionsFile:
-        lines = definitionsFile.readlines()
-        lineNumber = 1
-        for line in lines[1:]:
-            if not ":" in line:
-                continue
-            try:
-                identifier, value = line.split(':', 1)
-                definitions[identifier.strip()] = value.strip('" \n')
-            except ValueError:
-                print ("%d -> %s" % (lineNumber, line))
-            lineNumber += 1
-   
-    return definitions
 
-#Tries both cp1252 (ANSI) and UTF-8 encoding for reading files
-def findEncode(name):
-    functionOutput = ''
-
-    encodes = [ 'utf-8', 'cp1252' ]
-    for e in encodes:
-        try:
-            fh = open('%s/%s/%s' % (path, folder, name), 'r', encoding=e)
-            fh.readlines()
-            fh.seek(0)
-        except UnicodeDecodeError:
-            print('error on %s, trying another encoding' % e)
-        else:
-            print('encoding was %s' % e)
-            functionOutput = e 
-            return functionOutput
-
-#Splits the file at every bracket to ensure proper parsing
-def structureFile(name):
-    functionOutput = []
-    encode = findEncode(name)
-
-    with open('%s/%s/%s' % (path, folder, name), encoding=encode) as file:
-        for line in file:
-            if "#" in line:
-                line = line.split("#")[0]
-            if line == "":
-                continue
-            line = line.replace("{", "{\n").replace("}", "\n}").strip() #Splits line at brackets
-            if line == "":
-                continue
-            if "=" in line:
-                count = line.count("=")
-                if count > 1:
-                    for values in range(count):
-                        line = re.sub("(=[\s]*[\w\.]*) ([\w\.]*[\s]*=)", "\g<1>\n\g<2>", line) #Splits lines with more than one statement in two
-            if "\n" in line:
-                parts = line.split("\n")
-                for p in parts:
-                    functionOutput.append(p)
-            else:
-                functionOutput.append(line)
-    return functionOutput
- 
-#Determines the current level of nesting
-def nestingCheck(line):
-    global nesting
-    global nestingIncrement
-    nestingIncrement = 0
-    #Thanks to file restructuring, it is impossible for there to be multiple brackets on a line
-    if "{" in line:
-        nesting += 1
-        nestingIncrement = 1
-    elif "}" in line:
-        nesting -= 1
-        nestingIncrement = -1
- 
 def negationCheck(negative, line, negativeNesting):
     #Negation via NOT
     if "NOT" in line:
@@ -314,15 +231,6 @@ def formatLine(command, value, negative, random_list):
     line = statementLookup(line, statements, command, value)
     return line, negative
 
-def getValues(line):
-    line = line.split("=")
-    line[0] = line[0].strip()
-    try: #Checks if the command has a value
-        line[1] = line[1].strip().strip('{}"')
-        return line[0], line[1]
-    except IndexError:
-        return line[0], ""
- 
 def valueLookup(value, command):
     if value == "":
         return value, "other"
@@ -416,67 +324,68 @@ def output(line, negative): #Outputs line to a temp variable. Written to output 
     if specificFile != "no":
         print(line)
     outputText += line + "\n"
- 
-import cProfile, pstats
-pr = cProfile.Profile()
-pr.enable()
- 
-import time #Used for timing the parser
-start = time.clock()
-import re #Needed for various string handling
-import os #Used to grab the list of files
-settings = readStatements("settings")
-path = settings["path"].replace('\\', "")
-folder = settings["folder"]
-specificFile = settings["file"]
-if folder == "decisions":
-    nesting, nestingIncrement = 0, 0
-elif folder == "missions" or folder == "events":
-    nesting, nestingIncrement = 1, 0 #One less level of irrelevant nesting
- 
-#Dictionaries of known statements
-special = readStatements("statements/special")
-statements = readStatements("statements/statements")
-exceptions = readStatements("statements/exceptions")
- 
-try:
-    #Dictionaries of relevant values
-    provinces = readDefinitions("prov_names")
-    countries = readDefinitions("countries")
-    lookup = readDefinitions("EU4")
-    lookup.update(readDefinitions("text"))
-    lookup.update(readDefinitions("opinions"))
-    lookup.update(readDefinitions("powers_and_ideas"))
-    lookup.update(readDefinitions("decisions"))
-    lookup.update(readDefinitions("modifers"))
-    lookup.update(readDefinitions("muslim_dlc"))
-    lookup.update(readDefinitions("Purple_Phoenix"))
-    lookup.update(readDefinitions("core"))
-    lookup.update(readDefinitions("missions"))
-    lookup.update(readDefinitions("diplomacy"))
-    lookup.update(readDefinitions("flavor_events"))
-    lookup.update(readDefinitions("USA_dlc"))
-    events = readDefinitions("generic_events")
-    events.update(readDefinitions("flavor_events"))
-    events.update(readDefinitions("EU4"))
-    events.update(readDefinitions("muslim_dlc"))
-    events.update(readDefinitions("Purple_Phoenix"))
-    events.update(readDefinitions("USA_dlc"))
-    with open(path+"/common/event_modifiers/00_event_modifiers.txt") as f:
-        modifiers = f.readlines()
-   
-    if specificFile == "no":
-        for fileName in os.listdir("%s/%s" % (path, folder)):
-            print("Parsing file %s" % fileName)
+
+if __name__ == "__main__":
+    import cProfile, pstats
+    pr = cProfile.Profile()
+    pr.enable()
+
+    import time #Used for timing the parser
+    start = time.clock()
+    import re #Needed for various string handling
+    import os #Used to grab the list of files
+    settings = readStatements("settings")
+    path = settings["path"].replace('\\', "")
+    folder = settings["folder"]
+    specificFile = settings["file"]
+    if folder == "decisions":
+        nesting, nestingIncrement = 0, 0
+    elif folder == "missions" or folder == "events":
+        nesting, nestingIncrement = 1, 0 #One less level of irrelevant nesting
+
+    #Dictionaries of known statements
+    special = readStatements("statements/special")
+    statements = readStatements("statements/statements")
+    exceptions = readStatements("statements/exceptions")
+
+    try:
+        #Dictionaries of relevant values
+        provinces = readDefinitions(path, "prov_names")
+        countries = readDefinitions(path, "countries")
+        lookup = readDefinitions(path, "EU4")
+        lookup.update(readDefinitions(path, "text"))
+        lookup.update(readDefinitions(path, "opinions"))
+        lookup.update(readDefinitions(path, "powers_and_ideas"))
+        lookup.update(readDefinitions(path, "decisions"))
+        lookup.update(readDefinitions(path, "modifers"))
+        lookup.update(readDefinitions(path, "muslim_dlc"))
+        lookup.update(readDefinitions(path, "Purple_Phoenix"))
+        lookup.update(readDefinitions(path, "core"))
+        lookup.update(readDefinitions(path, "missions"))
+        lookup.update(readDefinitions(path, "diplomacy"))
+        lookup.update(readDefinitions(path, "flavor_events"))
+        lookup.update(readDefinitions(path, "USA_dlc"))
+        events = readDefinitions(path, "generic_events")
+        events.update(readDefinitions(path, "flavor_events"))
+        events.update(readDefinitions(path, "EU4"))
+        events.update(readDefinitions(path, "muslim_dlc"))
+        events.update(readDefinitions(path, "Purple_Phoenix"))
+        events.update(readDefinitions(path, "USA_dlc"))
+        with open(path+"/common/event_modifiers/00_event_modifiers.txt") as f:
+            modifiers = f.readlines()
+
+        if specificFile == "no":
+            for fileName in os.listdir("%s/%s" % (path, folder)):
+                print("Parsing file %s" % fileName)
+                main(fileName)
+        else:
+            fileName = specificFile+".txt"
             main(fileName)
-    else:
-        fileName = specificFile+".txt"
-        main(fileName)
-except FileNotFoundError:
-    print("File not found error: Make sure you've set the file path in settings.txt")
-elapsed = time.clock() - start
-print("Parsing the files took %.3f seconds" %elapsed)
-pr.disable()
-sortby = 'tottime'
-ps = pstats.Stats(pr).sort_stats(sortby)
-ps.print_stats()
+    except FileNotFoundError:
+        print("File not found error: Make sure you've set the file path in settings.txt")
+    elapsed = time.clock() - start
+    print("Parsing the files took %.3f seconds" %elapsed)
+    pr.disable()
+    sortby = 'tottime'
+    ps = pstats.Stats(pr).sort_stats(sortby)
+    ps.print_stats()
